move some headers
