+++ best/container/row.h
@@ -22,11 +22,11 @@
 
 #include "best/base/fwd.h"
 #include "best/base/ord.h"
+#include "best/base/tags.h"
 #include "best/container/internal/row.h"
 #include "best/container/object.h"
 #include "best/meta/empty.h"
 #include "best/meta/init.h"
-#include "best/meta/tags.h"
 #include "best/meta/taxonomy.h"
 #include "best/meta/tlist.h"
