+++ best/text/str.h
@@ -24,10 +24,10 @@
 
 #include <cstddef>
 
+#include "best/base/guard.h"
 #include "best/container/span.h"
 #include "best/math/overflow.h"
 #include "best/memory/bytes.h"
-#include "best/meta/guard.h"
 #include "best/meta/taxonomy.h"
 #include "best/text/encoding.h"
 #include "best/text/rune.h"
